Move freeze command to ModerationCommands
